Split up Spar.Data
