+++ services/spar/src/Spar/Sem/DefaultSsoCode/Cassandra.hs
@@ -22,8 +22,9 @@ module Spar.Sem.DefaultSsoCode.Cassandra where
 import Cassandra
 import Imports
 import Polysemy
-import qualified Spar.Data as Data
 import Spar.Sem.DefaultSsoCode
+import qualified SAML2.WebSSO.Types as SAML
+import Spar.Data.Instances ()
 
 defaultSsoCodeToCassandra ::
   forall m r a.
@@ -33,6 +34,44 @@ defaultSsoCodeToCassandra ::
 defaultSsoCodeToCassandra =
   interpret $
     embed @m . \case
-      Get -> Data.getDefaultSsoCode
-      Store ip -> Data.storeDefaultSsoCode ip
-      Delete -> Data.deleteDefaultSsoCode
+      Get -> getDefaultSsoCode
+      Store ip -> storeDefaultSsoCode ip
+      Delete -> deleteDefaultSsoCode
+
+-- It's important to maintain two invariants:
+-- 1) whenever there is a default code, it must also exist in the idp table
+-- 2) there can always only be one default SSO code selected
+
+getDefaultSsoCode ::
+  (HasCallStack, MonadClient m) =>
+  m (Maybe SAML.IdPId)
+getDefaultSsoCode =
+  runIdentity
+    <$$> (retry x1 . query1 sel $ params LocalQuorum ())
+  where
+    sel :: PrepQuery R () (Identity SAML.IdPId)
+    sel = "SELECT idp FROM default_idp WHERE partition_key_always_default = 'default' ORDER BY idp LIMIT 1"
+
+storeDefaultSsoCode ::
+  (HasCallStack, MonadClient m) =>
+  SAML.IdPId ->
+  m ()
+storeDefaultSsoCode idpId = do
+  -- there is a race condition here which means there could potentially be more
+  -- than one entry (violating invariant 2).
+  -- However, the SELECT query will deterministally pick one of them due to the
+  -- `ORDER BY` clause. The others will get removed by `deleteDefaultSsoCode`
+  -- the next time this function is called (as it removes all entries).
+  deleteDefaultSsoCode
+  retry x5 . write ins $ params LocalQuorum (Identity idpId)
+  where
+    ins :: PrepQuery W (Identity SAML.IdPId) ()
+    ins = "INSERT INTO default_idp (partition_key_always_default, idp) VALUES ('default', ?)"
+
+deleteDefaultSsoCode ::
+  (HasCallStack, MonadClient m) =>
+  m ()
+deleteDefaultSsoCode = retry x5 . write del $ params LocalQuorum ()
+  where
+    del :: PrepQuery W () ()
+    del = "DELETE FROM default_idp WHERE partition_key_always_default = 'default'"
