+++ services/spar/src/Spar/Sem/IdPRawMetadataStore/Cassandra.hs
@@ -17,11 +17,14 @@
 
 module Spar.Sem.IdPRawMetadataStore.Cassandra where
 
-import Cassandra
+import Cassandra as Cas
+import Control.Lens
+import Data.String.Conversions
 import Imports
+import qualified SAML2.WebSSO as SAML
 import Polysemy
-import qualified Spar.Data as Data
 import Spar.Sem.IdPRawMetadataStore
+import Spar.Data.Instances ()
 
 idpRawMetadataStoreToCassandra ::
   forall m r a.
@@ -31,6 +34,36 @@ idpRawMetadataStoreToCassandra ::
 idpRawMetadataStoreToCassandra =
   interpret $
     embed @m . \case
-      Store i t -> Data.storeIdPRawMetadata i t
-      Get i -> Data.getIdPRawMetadata i
-      Delete i -> Data.deleteIdPRawMetadata i
+      Store i t -> storeIdPRawMetadata i t
+      Get i -> getIdPRawMetadata i
+      Delete i -> deleteIdPRawMetadata i
+
+storeIdPRawMetadata ::
+  (HasCallStack, MonadClient m) =>
+  SAML.IdPId ->
+  ST ->
+  m ()
+storeIdPRawMetadata idp raw = retry x5 . write ins $ params LocalQuorum (idp, raw)
+  where
+    ins :: PrepQuery W (SAML.IdPId, ST) ()
+    ins = "INSERT INTO idp_raw_metadata (id, metadata) VALUES (?, ?)"
+
+getIdPRawMetadata ::
+  (HasCallStack, MonadClient m) =>
+  SAML.IdPId ->
+  m (Maybe ST)
+getIdPRawMetadata idp =
+  runIdentity
+    <$$> (retry x1 . query1 sel $ params LocalQuorum (Identity idp))
+  where
+    sel :: PrepQuery R (Identity SAML.IdPId) (Identity ST)
+    sel = "SELECT metadata FROM idp_raw_metadata WHERE id = ?"
+
+deleteIdPRawMetadata ::
+  (HasCallStack, MonadClient m) =>
+  SAML.IdPId ->
+  m ()
+deleteIdPRawMetadata idp = retry x5 . write del $ params LocalQuorum (Identity idp)
+  where
+    del :: PrepQuery W (Identity SAML.IdPId) ()
+    del = "DELETE FROM idp_raw_metadata WHERE id = ?"
