+++ services/spar/src/Spar/Sem/ScimExternalIdStore/Cassandra.hs
@@ -20,8 +20,10 @@ module Spar.Sem.ScimExternalIdStore.Cassandra where
 import Cassandra
 import Imports
 import Polysemy
-import qualified Spar.Data as Data
-import Spar.Sem.ScimExternalIdStore
+import Spar.Sem.ScimExternalIdStore (ScimExternalIdStore(..))
+import Brig.Types.Common (Email, fromEmail)
+import Data.Id
+
 
 scimExternalIdStoreToCassandra ::
   forall m r a.
@@ -31,6 +33,33 @@ scimExternalIdStoreToCassandra ::
 scimExternalIdStoreToCassandra =
   interpret $
     embed @m . \case
-      Insert tid em uid -> Data.insertScimExternalId tid em uid
-      Lookup tid em -> Data.lookupScimExternalId tid em
-      Delete tid em -> Data.deleteScimExternalId tid em
+      Insert tid em uid -> insertScimExternalId tid em uid
+      Lookup tid em -> lookupScimExternalId tid em
+      Delete tid em -> deleteScimExternalId tid em
+
+-- | If a scim externalId does not have an associated saml idp issuer, it cannot be stored in
+-- table @spar.user@.  In those cases, and only in those cases, we store the mapping to
+-- 'UserId' here.  (Note that since there is no associated IdP, the externalId is required to
+-- be an email address, so we enforce that in the type signature, even though we only use it
+-- as a 'Text'.)
+insertScimExternalId :: (HasCallStack, MonadClient m) => TeamId -> Email -> UserId -> m ()
+insertScimExternalId tid (fromEmail -> email) uid =
+  retry x5 . write insert $ params LocalQuorum (tid, email, uid)
+  where
+    insert :: PrepQuery W (TeamId, Text, UserId) ()
+    insert = "INSERT INTO scim_external (team, external_id, user) VALUES (?, ?, ?)"
+
+-- | The inverse of 'insertScimExternalId'.
+lookupScimExternalId :: (HasCallStack, MonadClient m) => TeamId -> Email -> m (Maybe UserId)
+lookupScimExternalId tid (fromEmail -> email) = runIdentity <$$> (retry x1 . query1 sel $ params LocalQuorum (tid, email))
+  where
+    sel :: PrepQuery R (TeamId, Text) (Identity UserId)
+    sel = "SELECT user FROM scim_external WHERE team = ? and external_id = ?"
+
+-- | The other inverse of 'insertScimExternalId' :).
+deleteScimExternalId :: (HasCallStack, MonadClient m) => TeamId -> Email -> m ()
+deleteScimExternalId tid (fromEmail -> email) =
+  retry x5 . write delete $ params LocalQuorum (tid, email)
+  where
+    delete :: PrepQuery W (TeamId, Text) ()
+    delete = "DELETE FROM scim_external WHERE team = ? and external_id = ?"
